# openapi.yaml
openapi: 3.0.3
info:
  title: DRP · Domestic Resource Planning API
  description: >
    API REST del core de DRP (Domestic Resource Planning).


    Este documento es la **fuente de verdad** de la API, no una ilustración
    de ella (ver ADR-007): se edita antes que el código que lo implementa, el
    frontend genera de él sus tipos y su cliente, y el backend se verifica
    contra él con las pruebas de contrato. Ver readme.md del proyecto para el
    contexto completo — modelo de dominio, event bus y roadmap.


    Dos convenciones transversales. Los **errores de forma** responden 400 con
    código `VALIDATION_ERROR`; los de **regla de negocio**, 409 con el código
    concreto. Y todas las colecciones se devuelven **paginadas**, con la misma
    envoltura `{ items, page, size, total }`.
  version: "0.3.0-fase2"
  license:
    name: Ver LICENSE del repositorio
servers:
  - url: http://localhost:8080/api/v1
    description: Desarrollo local, levantado por compose.yaml
  - url: https://api.drp.example/api/v1
    description: Entorno ilustrativo (el real se fija al elegir el VPS, al cerrar la Fase 1)
...
security:
  - bearerAuth: []
paths:
...
  /modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: Listar el catálogo de módulos con el estado del hogar
      description: >
        Devuelve el catálogo COMPLETO de lo que está desplegado, cada entrada
        con lo que este hogar ha decidido sobre ella. No se recorta por rol: ver
        que un módulo existe no es poder encenderlo --eso lo corta la
        autorización de las otras dos operaciones-- y quien no administra
        necesita saber que hay para poder pedirlo. Devolver dos catálogos según
        quién pregunte dejaría además al cliente sin poder distinguir "apagado"
        de "no me lo enseñan".


        La ausencia de decisión es INACTIVE: un hogar recién creado no tiene
        ninguna fila y todos sus módulos salen apagados. El cliente la resuelve
        UNA VEZ al arrancar la sesión, no una por pantalla.
      responses:
        "200":
          description: Catálogo de módulos con el estado del hogar
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModulePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /modules/{key}/activation:
    parameters:
      - name: key
        in: path
        required: true
        description: Clave del módulo, en UPPER_SNAKE_CASE
        schema: { type: string, pattern: "^[A-Z][A-Z0-9_]*$", example: SUPPLIERS }
    post:
      operationId: activateModule
      tags: [modules]
      summary: Activar un módulo en el hogar
      description: >
        Solo HOUSEHOLD_ADMIN. Enciende el módulo y lo SIEMBRA desde el estado
        actual del core -- no reproduciendo eventos: la entrega del bus es
        at-least-once y en memoria, así que un módulo activado hoy no vio el
        AssetCreated de hace un mes y no hay dónde ir a buscarlo.


        Es idempotente: activar lo ya activo devuelve el mismo estado y NO
        vuelve a sembrar. La siembra corre dentro de la misma transacción, así
        que un fallo suyo deja el módulo apagado en lugar de encendido a medias.
      responses:
        "200":
          description: Módulo activo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HouseholdModule" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/ServerError" }
    delete:
      operationId: deactivateModule
      tags: [modules]
      summary: Desactivar un módulo en el hogar
      description: >
        Solo HOUSEHOLD_ADMIN. CONSERVA LOS DATOS Y LOS OCULTA: el módulo deja de
        publicar sus rutas --responden 403 MODULE_INACTIVE--, deja de atender
        eventos y desaparece de la navegación, y sus filas siguen donde estaban.
        Volver a activarlo las devuelve tal cual.


        Se descartó borrar al desactivar, que es destructivo e irreversible y
        obligaría a definir el borrado en cascada de cuatro módulos antes de
        escribir el primero; y se descartó bloquear la desactivación cuando hay
        datos, que convierte la activación en casi irreversible y castiga justo
        al hogar que probó el módulo.
      responses:
        "200":
          description: Módulo inactivo. Sus datos siguen guardados.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HouseholdModule" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/ServerError" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT de usuarios del hogar. Claims -- sub (identityId), memberId
        (pertenencia en el hogar), householdId, role y sid. La distincion entre
        sub y memberId importa: todo lo que el dominio llama "usuario"
        --propietario de un asset, prestador, receptor, createdBy-- apunta a la
        pertenencia, mientras que los refresh tokens cuelgan de la identidad.


        sid identifica cual de las sesiones de esa persona es esta: es el
        identificador del refresh token con el que se emitio. Existe porque
        POST /auth/password revoca las DEMAS sesiones y conserva la que esta en
        uso, y para conservarla hay que saber cual es. La alternativa era pedir
        el refresh token en el cuerpo de esa operacion, que obligaria al cliente
        a mandar una credencial de vida larga donde no hace falta. No es un
        secreto: identifica una fila cuyo valor real esta hasheado.
    loanToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Credencial de un prestador o receptor **externo**, sin cuenta en DRP.
        Es un JWT firmado **sin sub de usuario**, con claims loanId y role
        (LENDER | BORROWER), que viaja como enlace por correo al iniciar el
        prestamo.


        Su alcance son exactamente dos operaciones -- GET /loans/{id} de ESE
        prestamo y POST /loans/{id}/return -- y ninguna otra cosa del hogar.
        Ni el listado, ni otro prestamo, ni los assets.


        Vale a la vez la firma y la fila: su hash se guarda en
        loan_access_tokens, y verificarlo exige firma valida **y** fila viva.
        Las dos capas hacen cosas distintas y por eso estan las dos -- la firma
        lo hace infalsificable y deja rechazar lo caducado sin tocar la base de
        datos; la fila lo hace **revocable**, que un JWT por si solo no puede
...
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: No autorizado para esta operación
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Falta el token, ha caducado o no es válido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ServerError:
      description: >
        Error no previsto (INTERNAL_ERROR). El cuerpo nunca lleva detalle
        interno: la traza va al log, no a la respuesta.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      description: >
        Forma única de todo error de la API. `code` es lo que el cliente trata;
        `message` es texto para diagnóstico, no para mostrar tal cual.
      properties:
        code:
          type: string
          example: ASSET_LOCATION_CONFLICT
          description: >
            Enumerado a propósito: el cliente decide qué hacer por el código, y
            un código que no esté aquí es un fallo del contrato, no un caso que
            el frontend deba adivinar.
          enum:
            - ALREADY_MEMBER
            - ARTICLE_DUPLICATE
            - ARTICLE_HAS_EXISTENCES
            - ARTICLE_UNIT_IMMUTABLE
            - ASSET_HAS_ACTIVE_LOAN
            - ASSET_HAS_CHILDREN
            - ASSET_LOCATION_CONFLICT
            - ASSET_QUANTITY_NEGATIVE
            - ASSET_QUANTITY_NOT_APPLICABLE
            - CATEGORY_DUPLICATE
            - CURRENT_PASSWORD_INVALID
            - DOCUMENT_CONTENT_INVALID
            - DOCUMENT_TARGET_INVALID
            - EMAIL_NOT_VERIFIED
            - EXISTENCE_ALREADY_IN_LOCATION
            - FILE_ALREADY_ATTACHED
            - FILE_IN_USE
            - FILE_TOO_LARGE
            - FILE_TYPE_NOT_ALLOWED
            - IDENTITY_ALREADY_MEMBER
            - INTAKE_QUANTITY_NOT_POSITIVE
            - INTERNAL_ERROR
            - INVITATION_ALREADY_PENDING
            - INVITATION_TOKEN_INVALID
            - LOAN_ALREADY_RETURNED
            - LOAN_ASSET_ALREADY_LENT
            - LOAN_ASSET_NOT_DURABLE
            - LOCATION_CYCLE
            - LOCATION_DUPLICATE
            - LOCATION_HAS_ASSETS
            - LOCATION_HAS_CHILDREN
            - MERGE_ARTICLE_MISMATCH
            - MERGE_ASSET_DEACTIVATED
            - MERGE_NOT_CONSUMABLE
            - MERGE_SAME_ASSET
            - MODULE_INACTIVE
            - RATE_LIMITED
            - RESET_TOKEN_INVALID
            - STORAGE_QUOTA_EXCEEDED
            - USER_LAST_ADMIN
            - VALIDATION_ERROR
            - VERIFICATION_TOKEN_INVALID
        message: { type: string }
        details:
          type: object
          description: Contexto adicional del error; su forma depende del código
    HouseholdModule:
      type: object
      required: [key, name, description, routePrefix, status]
      description: >
        Una entrada del catálogo con lo que este hogar ha decidido sobre ella.


        La clave, el nombre, la descripción y el prefijo de ruta NO se guardan
        por hogar: viven en el registro en código, que es donde está también lo
        que se despliega. Lo único que es del hogar es el estado y sus fechas.
      properties:
        key:
          type: string
          example: SUPPLIERS
          description: >
            Identificador del módulo, en inglés y en UPPER_SNAKE_CASE porque es
            un identificador y no un dato.
        name:
          type: string
          example: Proveedores y contactos de servicio
          description: Lo que ve el usuario. Es un dato, así que va en castellano.
        description: { type: string }
        routePrefix:
          type: string
          example: /api/v1/suppliers
          description: >
            Prefijo bajo el que el módulo publica sus operaciones. Todo lo que
            cuelga de él responde 403 MODULE_INACTIVE mientras esté apagado,
            exista o no todavía un controlador que lo atienda.
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
          description: >
            Sin decisión tomada el estado es INACTIVE. Para el cliente son el
            mismo caso, así que la respuesta no distingue "nunca se encendió" de
            "se apagó".
        activatedAt:
          type: string
          format: date-time
          nullable: true
          description: Cuándo se encendió por última vez
        deactivatedAt:
          type: string
          format: date-time
          nullable: true
          description: Cuándo se apagó por última vez
    PageMeta:
      type: object
      required: [page, size, total]
      description: >
        Metadatos de paginación, comunes a todas las colecciones. Son las mismas
        en las doce sin excepción por tamaño esperado: una sola forma que
        aprender en el cliente, y ninguna migración el día que una lista que se
        creía pequeña deje de serlo.
      properties:
        page: { type: integer, minimum: 0, example: 0, description: "Página devuelta, empezando en 0" }
        size: { type: integer, minimum: 1, example: 50, description: Tamaño de página aplicado }
        total: { type: integer, format: int64, example: 137, description: Total de elementos que cumplen el filtro }
...
    ModulePage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/HouseholdModule" }
...